requirements.txt
deep_translator==1.11.4 # 1.11.0
Django==4.2.9 
django-picklefield==3.1
ffmpeg==1.4
ffmpeg-python==0.2.0
google-cloud-texttospeech==2.14.2
googletrans==3.0.0
gTTS==2.3.2
librosa>=0.10.1
moviepy>=1.0.3
# av==11.0.0
mysqlclient==2.1.1
nltk==3.8.1
# numpy==1.26.1 # 1.21.6
openai-whisper==20230918
protobuf #>=3.6.1
pydub==0.25.1
pyffmpeg==2.4.2.18.1
scipy #>=1.11.3 # 1.10.1
six==1.16.0 # 1.16.0
social-auth-app-django==5.4.0
soundfile==0.12.1
# stripe>=5.0.0
# suno-bark==0.0.1a0
transformers #>=4.33.2 # 4.33.2
TTS #>=0.19.1 # 0.19.0
ttsmms==0.7
torch==2.0.1 # 2.0.1
